Refactor runTab actions
