+++ libs/remix-ui/run-tab/src/lib/constants/index.ts
@@ -44,3 +44,4 @@ export const RESET_STATE = 'RESET_STATE'
 export const ADD_DEPLOY_OPTION = 'ADD_DEPLOY_OPTION'
 export const REMOVE_DEPLOY_OPTION = 'REMOVE_DEPLOY_OPTION'
 export const SET_DEPLOY_OPTIONS = 'SET_DEPLOY_OPTIONS'
+export const SET_CURRENT_CONTRACT = 'SET_CURRENT_CONTRACT'
