+++ libs/remix-ui/run-tab/src/lib/run-tab.tsx
@@ -11,22 +11,22 @@ import { Modal, Network, RunTabProps, Tx } from './types'
 import { ContractData } from '@remix-project/core-plugin'
 import { runTabInitialState, runTabReducer } from './reducers/runTab'
 import {
-  initRunTab, setAccount,
-  setUnit, setGasFee,
-  setExecutionContext, setWeb3Endpoint,
-  clearPopUp, createNewBlockchainAccount,
-  setPassphrasePrompt, setMatchPassphrasePrompt,
-  signMessageWithAddress, getSelectedContract,
-  createInstance, setSendTransactionValue,
-  updateBaseFeePerGas, updateConfirmSettings,
-  updateGasPrice, updateGasPriceStatus,
-  updateMaxFee, updateMaxPriorityFee,
-  updateTxFeeContent, clearInstances,
-  removeInstance, getContext,
-  runTransactions, loadAddress,
-  storeScenario, runCurrentScenario,
-  updateScenarioPath, getFuncABIInputs,
-  setNetworkNameFromProvider
+  initRunTab, setAccountAddress,
+  setUnitValue, setGasFeeAmount,
+  setExecutionEnvironment,
+  hideToaster, createNewAddress,
+  setPassphraseModal, setMatchPassphraseModal,
+  signMessage, fetchSelectedContract,
+  createNewInstance, setSendValue,
+  setBaseFeePerGas, setConfirmSettings,
+  setGasPrice, setGasPriceStatus,
+  setMaxFee, setMaxPriorityFee,
+  setTxFeeContent, removeInstances,
+  removeSingleInstance, getExecutionContext,
+  executeTransactions, loadFromAddress,
+  storeNewScenario, runScenario,
+  setScenarioPath, getFuncABIValues,
+  setNetworkName
 } from './actions'
 import './css/run-tab.css'
 import { PublishToStorage } from '@remix-ui/publish-to-storage'
@@ -134,7 +134,7 @@ export function RunTabUI (props: RunTabProps) {
 
   const handleToaster = () => {
     setFocusToaster('')
-    clearPopUp()
+    hideToaster()
   }
 
   const toast = (toasterMsg: string) => {
@@ -171,11 +171,11 @@ export function RunTabUI (props: RunTabProps) {
   }
 
   const passphrasePrompt = (message: string) => {
-    return <PassphrasePrompt message={message} setPassphrase={setPassphrasePrompt} defaultValue={runTab.passphrase} />
+    return <PassphrasePrompt message={message} setPassphrase={setPassphraseModal} defaultValue={runTab.passphrase} />
   }
 
   const scenarioPrompt = (message: string, defaultValue) => {
-    return <ScenarioPrompt message={message} setScenarioPath={updateScenarioPath} defaultValue={defaultValue} />
+    return <ScenarioPrompt message={message} setScenarioPath={setScenarioPath} defaultValue={defaultValue} />
   }
 
   const mainnetPrompt = (tx: Tx, network: Network, amount: string, gasEstimation: string, gasFees: (maxFee: string, cb: (txFeeText: string, priceStatus: boolean) => void) => void, determineGasPrice: (cb: (txFeeText: string, gasPriceValue: string, gasPriceStatus: boolean) => void) => void) => {
@@ -186,13 +186,13 @@ export function RunTabUI (props: RunTabProps) {
       amount={amount}
       gasEstimation={gasEstimation}
       setNewGasPrice={gasFees}
-      updateBaseFeePerGas={updateBaseFeePerGas}
-      updateConfirmSettings={updateConfirmSettings}
-      updateGasPrice={updateGasPrice}
-      updateGasPriceStatus={updateGasPriceStatus}
-      updateMaxFee={updateMaxFee}
-      updateMaxPriorityFee={updateMaxPriorityFee}
-      setTxFeeContent={updateTxFeeContent}
+      updateBaseFeePerGas={setBaseFeePerGas}
+      updateConfirmSettings={setConfirmSettings}
+      updateGasPrice={setGasPrice}
+      updateGasPriceStatus={setGasPriceStatus}
+      updateMaxFee={setMaxFee}
+      updateMaxPriorityFee={setMaxPriorityFee}
+      setTxFeeContent={setTxFeeContent}
       txFeeContent={runTab.txFeeContent}
       maxFee={runTab.maxFee}
       maxPriorityFee={runTab.maxPriorityFee}
@@ -207,33 +207,32 @@ export function RunTabUI (props: RunTabProps) {
             networkName={runTab.networkName}
             personalMode={runTab.personalMode}
             selectExEnv={runTab.selectExEnv}
-            setWeb3Endpoint={setWeb3Endpoint}
             accounts={runTab.accounts}
-            setAccount={setAccount}
-            setUnit={setUnit}
+            setAccount={setAccountAddress}
+            setUnit={setUnitValue}
             sendValue={runTab.sendValue}
-            setSendValue={setSendTransactionValue}
+            setSendValue={setSendValue}
             sendUnit={runTab.sendUnit}
             gasLimit={runTab.gasLimit}
-            setGasFee={setGasFee}
+            setGasFee={setGasFeeAmount}
             providers={runTab.providers}
-            setExecutionContext={setExecutionContext}
-            createNewBlockchainAccount={createNewBlockchainAccount}
-            setPassphrase={setPassphrasePrompt}
-            setMatchPassphrase={setMatchPassphrasePrompt}
+            setExecutionContext={setExecutionEnvironment}
+            createNewBlockchainAccount={createNewAddress}
+            setPassphrase={setPassphraseModal}
+            setMatchPassphrase={setMatchPassphraseModal}
             modal={modal}
             tooltip={toast}
-            signMessageWithAddress={signMessageWithAddress}
+            signMessageWithAddress={signMessage}
             passphrase={runTab.passphrase}
           />
           <ContractDropdownUI
             exEnvironment={runTab.selectExEnv}
             contracts={runTab.contracts}
-            getSelectedContract={getSelectedContract}
+            getSelectedContract={fetchSelectedContract}
             modal={modal}
             passphrase={runTab.passphrase}
-            setPassphrase={setPassphrasePrompt}
-            createInstance={createInstance}
+            setPassphrase={setPassphraseModal}
+            createInstance={createNewInstance}
             ipfsCheckedState={runTab.ipfsChecked}
             setIpfsCheckedState={setCheckIpfs}
             publishToStorage={publishToStorage}
@@ -242,9 +241,9 @@ export function RunTabUI (props: RunTabProps) {
             passphrasePrompt={passphrasePrompt}
             mainnetPrompt={mainnetPrompt}
             tooltip={toast}
-            loadAddress={loadAddress}
+            loadAddress={loadFromAddress}
             networkName={runTab.networkName}
-            setNetworkName={setNetworkNameFromProvider}
+            setNetworkName={setNetworkName}
             deployOptions={runTab.deployOptions}
           />
           <RecorderUI
@@ -252,23 +251,23 @@ export function RunTabUI (props: RunTabProps) {
             logBuilder={logBuilder}
             passphrasePrompt={passphrasePrompt}
             mainnetPrompt={mainnetPrompt}
-            storeScenario={storeScenario}
-            runCurrentScenario={runCurrentScenario}
+            storeScenario={storeNewScenario}
+            runCurrentScenario={runScenario}
             scenarioPrompt={scenarioPrompt}
             count={runTab.recorder.transactionCount}
           />
           <InstanceContainerUI
             instances={runTab.instances}
-            clearInstances={clearInstances}
-            removeInstance={removeInstance}
-            getContext={getContext}
+            clearInstances={removeInstances}
+            removeInstance={removeSingleInstance}
+            getContext={getExecutionContext}
             gasEstimationPrompt={gasEstimationPrompt}
             logBuilder={logBuilder}
             passphrasePrompt={passphrasePrompt}
             mainnetPrompt={mainnetPrompt}
-            runTransactions={runTransactions}
+            runTransactions={executeTransactions}
             sendValue={runTab.sendValue}
-            getFuncABIInputs={getFuncABIInputs}
+            getFuncABIInputs={getFuncABIValues}
           />
         </div>
       </div>
